+++ LoginPageTests/LogInRepositoryTests.cs
@@ -74,7 +74,7 @@ namespace LoginPageTests.Tests
             bool result = await _logInRepository.CreateAccount(model);
             Assert.IsTrue(result);
 
-            string query = "DELETE FROM Users WHERE Username = @username";
+            string query = "DELETE FROM Logs WHERE UserId=(SELECT TOP 1 UserId FROM Users WHERE Username=@username)";
 
             using SqlConnection connectionToDatabase = new SqlConnection(Config.GetInstance().DatabaseConnection);
             await connectionToDatabase.OpenAsync().ConfigureAwait(false);
@@ -82,7 +82,12 @@ namespace LoginPageTests.Tests
             using SqlCommand command = new SqlCommand(query, connectionToDatabase);
             command.Parameters.AddWithValue("@username", "nexuser");
 
+            query = "DELETE FROM Users WHERE Username=@username";
             int rowsAffected = await command.ExecuteNonQueryAsync().ConfigureAwait(false);
+
+            using SqlCommand command2 = new SqlCommand(query, connectionToDatabase);
+            command2.Parameters.AddWithValue("@username", "nexuser");
+            int rowsAffected2 = await command2.ExecuteNonQueryAsync().ConfigureAwait(false);
         }
 
         [TestMethod]
